requirements.txt
Django==3.0
django-phonenumber-field==1.1.0
djangorestframework==3.11.0